Remove cpulimit from couchdb commit validation
As the build slaves are quite big machines now, I think we can try
to remove the cpulimit code. This should make the build time shorter.
The reason it was there was for the dialyzer runs, when the PLT file
was created. This took so much resources, so that Jeknins killed the
job.

I would just dare to remove it and see if it now survives without
being killed by Jenkins. If not, we can always revert this commit.

Change-Id: Ie09617cf957697bf76fbeb1d60493993884c7a50
Reviewed-on: http://review.couchbase.org/39236
Reviewed-by: Philip Labee <philip@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Volker Mische <volker.mische@gmail.com>