CBD-1425:  fix up manifest generation/coordination
If we pass in the names of manifests ( e.g. branch-master.xml
+ override-master.xml), pass both to fetch-manifest.rb, specifying
VLT_DIR/current.xml as output.

If we pass in current.xml for manifest, download build-specific
manifest into VLT_DIR/current.xml, then pass this to fetch-manifest.rb
and blank as the override.  Output over-writes currently.xml in
the voltron dir.

The make command that follows expects VLT_DIR/current.xml

Change-Id: I92674bca6dbba4a6efaa18f3d16c93a83cfec69f
Reviewed-on: http://review.couchbase.org/40620
Reviewed-by: Philip Labee <philip@couchbase.com>
Tested-by: Philip Labee <philip@couchbase.com>