MB-12913: Use redhat-lsb-core instead of redhat-lsb
The `redhat-lsb` package (needed for the lsb_release command) brings
in a large number of dependancies. However we only actually need the
`redhat-lsb-core` package to run the command-line programs, so change
to this.

Note that the git compile command also needs updating as it implicitly
made use of packages installed by redhat-lsb. We now explicitly list
its required packages.

Change-Id: Iaed3b7f9111ebe64880341cb2e0197283b0f0d0f
Reviewed-on: http://review.couchbase.org/44464
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>