MB-12915: cbdeps for libstdc++ (C++11)
Create a build to package up libstdc++.so (from a C++11 compliant
install) into a cbdeps package.

Ubuntu12.04 doesn't support C++11 by default, so we need to install
(at build-time) a newer compiler (GCC 4.8.2) which does support
it. This also installs an updated libstdc++.so to support the new
C++11 features. Libstdc++.so is needed at runtime (i.e. as part of the
packaged up Couchbase), so we use the cbdeps mechanism to package and
then copy into <install>/lib as part of the build, and then be part of
the packaged Couchbase.

Note that of the supported platforms we only need to do this for
Ubuntu 12.04 - Ubuntu 14.04 already supports C++11 from the default
libstdc++, and CentOS 6 solves this problem by statically linking the
new C++11 functionality in the applicaiton; meaning the original
system libstdc++.so can still be used.

Change-Id: Ieaccc80d34fda3c92b1bd672148ac1d9e81486ea
Reviewed-on: http://review.couchbase.org/44485
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>