Use $ORIGIN for rpath in cbdeps ICU4c
ICU (and other cbdeps) must be built so they can be installed in an
arbitrary location, as they are used both in development builds and
"official" production releases - for example a development build this
location might be $HOME/couchbase/install/lib, but an officual
production release would be at /opt/couchbase/lib.

ICU differs from the other cbdeps in that it contains shared objects
which are transitively required by Couchbase executables -
i.e. libcouchstore.so requires libicuuc.so, which itself requires
libicudata.so. This adds some complexity when building ICU as a cbdep,
as liicuuc.so needs to be able to locate libicudata.so at runtime,
even when the original compile-time location of ICU is different to
it's installed location.

To achieve this we need to set the RPATH of libicuuc.so (and similar
ICU libraries) to $ORIGIN, to tell the dynamic linker to seach for
libicuuc's dependancies (e.g. libicudata.so) in the same directory
that libicuuc.so exists).

Change-Id: I2bb0c69841f0e5f7f83b46b47c5d97ad41242ce7
Reviewed-on: http://review.couchbase.org/45679
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Dave Rigby <daver@couchbase.com>