MB-9994: Fetch all projects with the same Change-Id
Instead of fetching a single project for commit validation, get the
Change-Id of the current commit to validate and check out all projects
that contain such a Change-Id.

This commits adds a script (allcommits.py) that gets the Git ref by
Change-Id. It returns the project name, the path and the ref of the
given change id. If there are several commits with the same Change-Id
it will return all of them.

The output can be piped into another script (fetch_project.sh) that
goes into the right directory (as specified in the repo manifest),
resets it and fetches the new source.

This approach has several advantages:

 - You don't need to set the project name or path manually. This way
   a lot of commit validation scripts could be consolidated into a
   single one. It also prevents breakage whenever a project path is
   changed in the manifest (it happened during the period where the
   build system transitioned to CMake. There some projects weren't
   any longer at the top-level, but within the `cmake` directory).
 - When there are changes that somehow depend on each other, the
   developer can simply assign the same Change-Id and the buildbot
   will be able to build and test them together.

Possible future improvement:

 - Check the commit history of the current change and also fetch all
   projects that have the same Change-Id as prior commits that are
   not yet merged. This would enable even chains of dependencies
   between projects.

Change-Id: Icbc4ea9032c472c25ac7bf7cc08befc8b0be1ca9
Reviewed-on: http://review.couchbase.org/45935
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Volker Mische <volker.mische@gmail.com>