commit-validation: Allow parallelism for tests to be specified seperately
Some projects (e.g. forestdb) cannot handle running their tests in
parallel, but it is desirable to still build the project in parallel
(for speed) even if we have to run the tests sequentially.

Allow the test parallelism to be overridden, using the
TEST_PARALLELISM env var. If not specified then the same value as
build parallelism (i.e. env var PARALLELISM; defaults to 8) is used.

Change-Id: Ie02bdba6638423c0909b88e1108befc9dacf1bc9
Reviewed-on: http://review.couchbase.org/49435
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>