single-project-gerrit: Fail build if any tests fail
When building Couchbase we want to run as many tests as we can (to
maximise the information presented to developers), but at the same
time we also want to ensure that *any* errors encountered are reported
back to Jenkins (i.e. the final exit status should be non-zero).

To achieve this, split the script into two stages:

1. An initial manditory stage where errors are immediately fatal. This
   is used for the actual building of Couchbase; which if it fails
   there is no point in trying to run tests.

2. An subsequent testing stage where any errors encountered are noted,
   but we defer reporting them until the end of the script.

Change-Id: I8aaf2ada1341dbe3458d49900527c71a2b0f41f9
Reviewed-on: http://review.couchbase.org/54448
Well-Formed: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Hari Kodungallur <hari.kodungallur@couchbase.com>