If nopkg option selected, exit out before any packaging step
The cleanup.rb is called to clean files from /opt/couchbase and remove
any files that are not needed for packaging.

For running unit tests, the best approach is to call this script
with nopkg option and then run make test. In which case, the nopkg
should leave all dependencies in /opt/couchbase and not clean them up

Change-Id: Ic8787d24dc4f487154f2a86eff7e9f95e7ab9720
Reviewed-on: http://review.couchbase.org/55224
Well-Formed: buildbot <build@couchbase.com>
Reviewed-by: Hari Kodungallur <hari.kodungallur@couchbase.com>
Tested-by: Hari Kodungallur <hari.kodungallur@couchbase.com>