disable log collection on macos
macos tests are failing, likely because cbserver doesn't work
on mavericks anymore. but don't want to disable it altogether
as 4.1.0 looks like it may be working (still to verify).
in any case, these logs are useless on macos at this point of
time. it is just adding up on storage on mega

Change-Id: Ifbb97206f028881dd0516cfe0203be3d4c19a7b9
Reviewed-on: http://review.couchbase.org/56950
Well-Formed: buildbot <build@couchbase.com>
Reviewed-by: Hari Kodungallur <hari.kodungallur@couchbase.com>
Tested-by: Hari Kodungallur <hari.kodungallur@couchbase.com>