CBD-1625: Include relative path shared library in core file archive
When analysing core files at the end of the run, also include shared
libraries which were loaded using a relative path. This requires that
we extract the working directory from the core file, and then change
to this directory before copying shared libraries to our archive.

Also ensure that debuginfo files are picked up if they are symlinks.

Change-Id: I2974988a6e489927ac54fe1794a13e0b829ef69f
Reviewed-on: http://review.couchbase.org/60838
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>