Always checkout the refspec given
In some edge cases the refspec given will not be valid
for the manifest we're using (i.e. uses the wrong branch).In this
case we should just use it anyway.

Also checks that at least one change has been found and adds further
error checking to fetch_project.sh and fetch_project.bat in case of
an empty line.

Change-Id: I220bec846c05fe353f57c6c60ac2cf41f4aefa05
Reviewed-on: http://review.couchbase.org/60494
Tested-by: Will Gardner <will.gardner@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>