CBD-1625: Fix incorrect detection of core file when none exist
The archive_core_files.sh script was called with a glob (/tmp/core.*),
however if no match is found then it passes the literal '/tmp/core.*'
to the script. This resulted in core files being 'detected' (and the
build failed) when none were present.

Fix by setting the 'nullglob' shell option to remove the glob and pass
nothing if no cores exist.

Change-Id: Ibb5257ec6c7cee684e73723f74b64ec550ea8c74
Reviewed-on: http://review.couchbase.org/61338
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>