command + instructions to setup jenkins slave on sanity test machine
in addition to a cluster we need to a jenkins slave that runs testrunner
and runs tests on the given set of IPs (cluster). This slave need not
be one of the machines in the cluster. If we use docker containers,
most of these things are already setup, but for VMs they are not.

Change-Id: Ib084ff08aa2815fe7a2f6eaeddd747480f79be52
Reviewed-on: http://review.couchbase.org/67135
Reviewed-by: Hari Kodungallur <hari.kodungallur@couchbase.com>
Tested-by: Hari Kodungallur <hari.kodungallur@couchbase.com>