CBD-1904 cv work for .net sdk and also infra setup
This is also may be tracked in SDKQE-789.

launch_instance.py: provisions 4 VMs for testing. The parameters
are hardcoded for now. The params being ami-id, sg-id (security
group), number of instances, and type of instance. If and when
needed, this can, of course, be changed to be command line args

setup_multi_node_cluster.sh: sets up a 4 node cluster with beer-sample,
default and defaul1 buckets installed.

updated sdk_win_instance to also have a "status" function. not used
now, but i think it will be needed (the UTC time reported by the API
seems one hour off though)

Change-Id: I170c7880e23bc03309f0859089ba732e0f8b6ae5
Reviewed-on: http://review.couchbase.org/67952
Reviewed-by: Hari Kodungallur <hari.kodungallur@couchbase.com>
Tested-by: Hari Kodungallur <hari.kodungallur@couchbase.com>