also monitor for any rogue test instances still running
probably we should force kill if test instances are running
build instance is reused and there is a chance some job could be using it
but test instance should never be reused, and so we should monitor those
instances separately and kill them as soon as we find them to be running
for more than, say, an hour. [Unless there is some edge case (jenkins goes
down or the job gets abnormally killed), the test instance should always
be terminted by the job after the test. The monitoring is for such edge
cases]

Change-Id: I484e9e5a8907c663b851ed303f25fe0daeef586a
Reviewed-on: http://review.couchbase.org/68687
Reviewed-by: Hari Kodungallur <hari.kodungallur@couchbase.com>
Tested-by: Hari Kodungallur <hari.kodungallur@couchbase.com>