Commit validation: Add support for running git-clang-format
This adds a new script. `single-project-clang-format` which runs
git-clang-format against the specified Gerrit patch, in a similar
way to the other commit-validation script.

This only runs against files which are already conforming to the
style format (or are new) so will allow for enforcing conformity
on a file-by-file basis.

This is a separate script so that it can be setup as a separate job
and trigger a `-1 code review` instead of `-1 verified`. Because it
is lightweight it can probably run on the cv-zz-lightweight slave.

Change-Id: I044e8e15abafdcd1ee6a1721f352038a52953a29
Reviewed-on: http://review.couchbase.org/70881
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Will Gardner <willg@rdner.io>