fetch_project: don't fail checkout if working dir not clean
When fetching a project for a given change, the checkout can fail if
the working directory is not clean - for example if a new file has
recently been added, but the previous checkout was older and didn't
include that file. For example:

    error: The following untracked working tree files would be overwritten by checkout:
    src/kv_bucket.h
    Please move or remove them before you can switch branches.
    Aborting

To solve this, perform a `git clean` before the checkout, removing any
files which are not owned by git.

Change-Id: I982cc44c768f6669e01d313efc4b8844a24ced5a
Reviewed-on: http://review.couchbase.org/71760
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>