Commit validation: remove CMakeCache.txt during clean
When switching between patches which are based on different parents,
we can encounter build failures if CMake options which were added in
one were not present in the other.

For example, a recent change to the default value of EP_USE_FORESTDB
can mean that the build can attempt to link forestdb.lib, even though
it hasn't been built.

To address this, remove CMake's cache file (CMakeCache.txt) as part of
the clean step - ensuring that any CMake variables are calculated from
scratch for each CV job.

Change-Id: I410def56081bd318c229030f0fa203f82d035d78
Reviewed-on: http://review.couchbase.org/71989
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Dave Rigby <daver@couchbase.com>