MB-24722: Use different filenames for unsigned builds
For macOS and Windows, we want to use different names for the files
before they are signed, to allow us to tell if a signing didn't occur.
We do this by adding '-unsigned' to the end of each filename (pre-suffix)
and then create the signed file with the extension left out.

Note that the signing process for Windows is handled in the Jenkins file
itself, so it not noted here.

Change-Id: I75546f897b1ce0533ca11dc909f272e73c636471
Reviewed-on: http://review.couchbase.org/79280
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>