Replace backticks with subshell and inline in conditional
Apparently using '-e' is hazardous to one's health; if any command in
the script fails, it will immediately abort and return a non-zero exit
code.  Seemingly using backticks to run a command causes this; however,
using a subshell doesn't, so use that instead and while we're at it,
inline the sub-command into the conditional it was being used for.

Change-Id: I462bcccd9c0e0fe505e8c270f32b972fe2047e3e
Reviewed-on: http://review.couchbase.org/79342
Reviewed-by: Kenneth Lareau <kenneth.lareau@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>