MB-24772: Handle signing name change on watson branch
With the name change for unsigned macOS builds, the watson branch needs
the same change as mainline (spock) to ensure '-unsigned' is at the end
of the filename for initial builds.  The server-linux-build.sh is the
only one that technically needs to be changed, but updated the scripts
codesignApple.sh and server_macos_build.sh for consistency/sanity reasons.

Change-Id: Ib2de3e833eefb67823beb8b4cacfdbadc27c5d78
Reviewed-on: http://review.couchbase.org/79398
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>