archive_core_files: support versioned 'environ' symbol
To determine the current working directory of the core, we need to
access the environment variable 'PWD' from the core file. To do this
we examine the memory at the `environ` symbol.

Newer toolchains use a versioned symbol - e.g. 'environ@GLIBC_2.2.5'
and as such we fail to read the environment.

Add support for a versioned `environ` symbol; also check that we
successfully read the environment from the core and print an error if
not.

Change-Id: I1b66769f8d5170e7caf69a8c81cbd9db7853b46f
Reviewed-on: http://review.couchbase.org/79885
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>