MB-24724: Handle unexpected exceptions (avoid 404)
This is a very brutish way to handle the restricted-branch-check program
throwing unexpected exceptions and not generating an HTML page for the
user (leading to a 404); the main logic of the program is wrapped by an
all-encompassing try/except that catches ALL exceptions and generates
an HTML page with the exception.

Future enhancements should remove the catch-all and look for specific
exceptions to handle retries when possible and the like, but this will
suffice for now.

Change-Id: I8cffddc72746e8ba1c5e7202aeac4a904aa3ed8f
Reviewed-on: http://review.couchbase.org/80211
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>