CBD-2419: Clean up script to do Jenkins backups.
Some basic fixes and clean up for the Jenkins backup script, which will
be run from each Jenkins server as a job to handle it's own backups on
a nightly basis.  Some modern bashisms were used during the clean up
(primarily to avoid needing to shell out to /bin/test).

Change-Id: I79074936a44edc4bf33fff4cf23bec585c93025f
Reviewed-on: http://review.couchbase.org/93129
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>