MB-29037: Revert Vulcan Centos7 buildslave to Centos 7.2
Centos 7.4 updated to OpenSSL 1.0.2k. While this doesn't directly affect
Server builds, it appears that it does affect cbdeps builds - in particular,
Erlang built on that slave has a hard dependency on OpenSSL 1.0.2. As
Server's platform support states all Centos 7 flavors are supported and
Centos 7.2 is certified, we must only link against OpenSSL 1.0.1. The most
reliable way to achieve that is to switch the buildslave back to Centos 7.2.

As Centos 7 is a rolling release, even just switching to the centos:7.2.1511
base image was not sufficient. We have to lock the yum repositories to
Centos' Vault for 7.2.1511 - very pleased that they offer those.

Change-Id: I8249afa950f720565b6c918b433f4a8932fede0c
Reviewed-on: http://review.couchbase.org/93342
Tested-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Kenneth Lareau <kenneth.lareau@couchbase.com>