CBD-2560: Clean up old buildboard / build database files
With the completed migration of all needed functionality from the old
buildboard and build database setups, quite a bit can be cleaned out
of the repository, as there is no plans to resurrect any of it.  This
includes some data in scripts that was for a very old web interface.

Change-Id: I6be835ddba468078ec3956b3b6a194cd1af5dc8d
Reviewed-on: http://review.couchbase.org/95843
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Kenneth Lareau <kenneth.lareau@couchbase.com>