archive_core_files.sh: Add last-chance method to locate pwd
If we have failed to determine the working directory of the core dump
with the first two (scientific) methods, make a final attempt by using
'strings' on the core dump. This could result in false positives, but
it seems better to try and potentially get the correct path than just
give up.

Change-Id: I20fd3a60395cdff4227758f65023112e19e592c0
Reviewed-on: http://review.couchbase.org/104839
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>