single-project-gerrit*: Don't print env vars
We currently print all environment variables set to the
single-project-gerrit logfiles.

This can cause problems if the commit message of the patch (exposed
via GERRIT_CHANGE_COMMIT_MESSAGE) contains strings which would
otherwise be detected as actual warnings - for example it's common for
a commit message to include an extrace of the warning/error which is
being fixed. By printing these to the log file, the warnings/error log
file scanner incorrectly reports these as actual errors in the build,
resulting in the build being failed.

To fix this problem, remove the printing of environment variables from
the single-project-gerrit* jobs. Note that the environment variable
are recorded by Jenkins anyway and can be viewed from
'injectedEnvVars/' URL suffix on the job, so we arn't loosing any
information here.

Change-Id: I60f46bc6b684b61e9fe19895beb91e24b0f0b65d
Reviewed-on: http://review.couchbase.org/110150
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>