Don't rebuild without -j8 for error checking
In the past, if there was a build failure, we re-ran "make" without "-j8" to
produce cleaner output for debugging the problem. However, due to the way
-j8 threading works, often that re-run would build (slowly) a lot of stuff
that didn't get attempted the first time before ultimately getting back to
the build error. This sometimes caused build jobs to take upwards of 45
minutes.

Change-Id: Ice5343d42d29bc000179d991598d37df4ec3cd3d
Reviewed-on: http://review.couchbase.org/111271
Reviewed-by: Kim Trang <kim.trang@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>