CBD-2986: Don't enable CB_PRODUCTION_BUILD in script
We want CB_PRODUCTION_BUILD to *only* be enabled by specific Jenkins jobs,
namely the production build jobs. It shouldn't be embedded in the build
scripts as those scripts are used elsewhere, such as in toy builds.

Analytics uses CB_PRODUCTION_BUILD to decide whether to download the JDK
or not, but we can also specify CB_DOWNLOAD_JAVA to enable that, so we have
the build scripts set that flag instead.

Change-Id: Ic12d78b5b5c0209cdc9a330066e176d7b5f500b5
Reviewed-on: http://review.couchbase.org/111676
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>