MB-37484: Fix scan-build so it uses clang options
couchstore changes related to this MB do not compile with the scan
build as there is a mixture of clang compiler but configured from
CouchbaseGccOptions.cmake. This means the use of the C++14 sized
delete operator will cause a compilation failure.

This change means that the scan build will configure using
CouchbaseClangOptions, and the C++14 sized delete will be enabled.

Change-Id: I9a943416c534ef80df0e5391255ea66921bb7f8e
Reviewed-on: http://review.couchbase.org/120885
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>