CBD-3754: Don't checkout other project in clang-format
The new patch_via_gerrit script fails if a project with a linked
change exists but has not been checked out. In the case of clang
format we do not check out EE repos so links to magma changes do
not work. The clang format job doesn't need to check out other
repos though as it should only ever run on the project being
tested and no code is built.

Change-Id: I23b14dc6740b2584acd4f0ce86b5d1a28d1f454e
Reviewed-on: http://review.couchbase.org/c/build/+/160294
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>