CBD-3754: Add checkout_dependencies to clang format CV
Currently the clang format CV is broken as it does not checkout the
latest pach, it just runs on the master branch. Ideally we would use
the newer patch_via_gerrit here but if we link with a magma patch
then the kv job fails as it only checks out the CE repos and
patch_via_gerrit errors upon not being able to find the magma repo.

Fix this by using the older checkout_dependencies to checkout the
KV (and other) repo(s).

Change-Id: Ic1caad0df6edd4308343aa448550fe1be1f7bc12
Reviewed-on: http://review.couchbase.org/c/build/+/160945
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>