CBD-5165: Propogate PARALLELISM to single-project-gerrit.bat
While the PARALLELISM env var is set in single-project-gerrit.bat for
Windows builds, it is not actually used when building (unlike
single-project-gerrit.sh for *ix builds).

Propagate it through, so we can limit build parallelism and hopefully
reduce / avoid instances of the CV machines running out of memory when
building with default parallelism.

Change-Id: I4866366dbf472751f310f2175e42d3e7fa8a3f7b
Reviewed-on: https://review.couchbase.org/c/build/+/181033
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>