Add TESTS_INCLUDE env var support to single project gerrit
So that we can specify a regex of tests that we actually want to run,
rather than a regex of tests that we don't want to run. Passes default
values to TESTS_EXCLUDE that matches nothing, and a default to
TESTS_INCLUDE that matches everything, to maintain the existing
behaviour.

Change-Id: Iefe002c6bd59186f206c1976c35de88bf21409ba
Reviewed-on: https://review.couchbase.org/c/build/+/199764
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>