Add TESTS_INCLUDE env var support to single project gerrit
Patch modified from https://review.couchbase.org/c/build/+/199764

The original patch was excluding a bunch of tests on windows. This
has been fixed by using "a^^" and ".*" for the exclude and include
vars.

Original commit message:

So that we can specify a regex of tests that we actually want to run,
rather than a regex of tests that we don't want to run. Passes default
values to TESTS_EXCLUDE that matches nothing, and a default to
TESTS_INCLUDE that matches everything, to maintain the existing
behaviour.

Author: Ben Huddleston <ben.huddleston@couchbase.com>
Change-Id: Idb1ffb8e733a9e910670898a4dee873336f51851
Reviewed-on: https://review.couchbase.org/c/build/+/201761
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>